Split indexer verticle to extract chunk functions
