+++ georocket-server/src/main/java/io/georocket/index/elasticsearch/ElasticsearchClient.java
@@ -35,11 +35,6 @@ public class ElasticsearchClient {
   private final String index;
   
   /**
-   * The type of objects stored in the index
-   */
-  private final String type;
-  
-  /**
    * The Vert.x instance
    */
   private final Vertx vertx;
@@ -55,13 +50,10 @@ public class ElasticsearchClient {
    * @param port the port on which Elasticsearch is listening for HTTP
    * requests (most likely 9200)
    * @param index the index to query against
-   * @param type the type of objects stored in the index
    * @param vertx a Vert.x instance
    */
-  public ElasticsearchClient(String host, int port, String index, String type,
-      Vertx vertx) {
+  public ElasticsearchClient(String host, int port, String index, Vertx vertx) {
     this.index = index;
-    this.type = type;
     this.vertx = vertx;
     
     HttpClientOptions clientOptions = new HttpClientOptions()
@@ -85,7 +77,7 @@ public class ElasticsearchClient {
    * @see #bulkResponseHasErrors(JsonObject)
    * @see #bulkResponseGetErrorMessage(JsonObject)
    */
-  public Observable<JsonObject> bulkInsert(Map<String, JsonObject> documents) {
+  public Observable<JsonObject> bulkInsert(String type, Map<String, JsonObject> documents) {
     String uri = "/" + index + "/" + type + "/_bulk";
     
     // prepare the whole body now because it's much faster to send
@@ -109,9 +101,9 @@ public class ElasticsearchClient {
    * @return an object containing the search hits and a scroll id that can
    * be passed to {@link #continueScroll(String, String)} to get more results
    */
-  public Observable<JsonObject> beginScroll(JsonObject query, int pageSize,
+  public Observable<JsonObject> beginScroll(String type, JsonObject query, int pageSize,
       String timeout) {
-    return beginScroll(query, null, pageSize, timeout);
+    return beginScroll(type, query, null, pageSize, timeout);
   }
   
   /**
@@ -127,9 +119,9 @@ public class ElasticsearchClient {
    * @return an object containing the search hits and a scroll id that can
    * be passed to {@link #continueScroll(String, String)} to get more results
    */
-  public Observable<JsonObject> beginScroll(JsonObject query, JsonObject postFilter,
+  public Observable<JsonObject> beginScroll(String type, JsonObject query, JsonObject postFilter,
       int pageSize, String timeout) {
-    return beginScroll(query, postFilter, null, pageSize, timeout);
+    return beginScroll(type, query, postFilter, null, pageSize, timeout);
   }
 
   /**
@@ -146,7 +138,7 @@ public class ElasticsearchClient {
    * @return an object containing the search hits and a scroll id that can
    * be passed to {@link #continueScroll(String, String)} to get more results
    */
-  public Observable<JsonObject> beginScroll(JsonObject query, JsonObject postFilter,
+  public Observable<JsonObject> beginScroll(String type, JsonObject query, JsonObject postFilter,
       JsonObject aggregation, int pageSize, String timeout) {
     String uri = "/" + index + "/" + type + "/_search";
     uri += "?scroll=" + timeout;
@@ -171,7 +163,7 @@ public class ElasticsearchClient {
   
   /**
    * Continue scrolling through search results. Call
-   * {@link #beginScroll(JsonObject, int, String)} to get a scroll id
+   * {@link #beginScroll(String, JsonObject, int, String)} to get a scroll id
    * @param scrollId the scroll id
    * @param timeout the time after which the scroll id becomes invalid
    * @return an object containing new search hits and possibly a new scroll id
@@ -193,7 +185,7 @@ public class ElasticsearchClient {
    * @see #bulkResponseHasErrors(JsonObject)
    * @see #bulkResponseGetErrorMessage(JsonObject)
    */
-  public Observable<JsonObject> bulkDelete(JsonArray ids) {
+  public Observable<JsonObject> bulkDelete(String type, JsonArray ids) {
     String uri = "/" + index + "/" + type + "/_bulk";
     
     // prepare the whole body now because it's much faster to send
@@ -222,7 +214,7 @@ public class ElasticsearchClient {
    * @return an observable emitting <code>true</code> if the type of
    * the index exists or <code>false</code> otherwise
    */
-  public Observable<Boolean> typeExists() {
+  public Observable<Boolean> typeExists(String type) {
     return exists("/" + index + "/" + type);
   }
 
@@ -244,12 +236,12 @@ public class ElasticsearchClient {
   }
   
   /**
-   * Create the index
+   * Create the index for given type and mapping.
    * @param mappings the mappings to set for the index
    * @return an observable emitting <code>true</code> if the index creation
    * was ackowledged by Elasticsearch, <code>false</code> otherwise
    */
-  public Observable<Boolean> createIndex(JsonObject mappings) {
+  public Observable<Boolean> createIndex(String type, JsonObject mappings) {
     String uri = "/" + index + "/_mapping/" + type;
 
     return performRequestRetry(HttpMethod.PUT, uri, mappings.encode())
@@ -343,8 +335,8 @@ public class ElasticsearchClient {
    * @param response the bulk response
    * @return <code>true</code> if the response has errors, <code>false</code>
    * otherwise
-   * @see #bulkInsert(Map)
-   * @see #bulkDelete(JsonArray)
+   * @see #bulkInsert(String, Map)
+   * @see #bulkDelete(String, JsonArray)
    * @see #bulkResponseGetErrorMessage(JsonObject)
    */
   public boolean bulkResponseHasErrors(JsonObject response) {
@@ -356,8 +348,8 @@ public class ElasticsearchClient {
    * @param response the response containing errors
    * @return the error message or null if the bulk response does not contain
    * errors
-   * @see #bulkInsert(Map)
-   * @see #bulkDelete(JsonArray)
+   * @see #bulkInsert(String, Map)
+   * @see #bulkDelete(String, JsonArray)
    * @see #bulkResponseHasErrors(JsonObject)
    */
   public String bulkResponseGetErrorMessage(JsonObject response) {
